Removed the Issue_.java generated class in remrem semantics.
Implemented code changes which avoids the generation of duplicate class
names.
